+++ src/app/[orgSlug]/study/[studyId]/_screens/execution-stage-guard.tsx
@@ -1,0 +1,38 @@
+import { isSubmittedStudy } from '@/schema/study'
+import { AlertNotFound } from '@/components/errors'
+import { currentExecutionStage } from '@/lib/study-job-status'
+import { latestSubmittedJobForStudy, type LatestJobForStudy } from '@/server/db/queries'
+import type { SelectedStudy } from '@/server/actions/study.actions'
+
+type GuardSuccess = {
+    job: LatestJobForStudy
+    stage: NonNullable<ReturnType<typeof currentExecutionStage>>
+}
+
+// Returns { job, stage } on success, or a ReactElement alert on failure.
+// Callers discriminate with `!('job' in result)` to catch all three failure paths.
+export async function guardExecutionStage(
+    study: SelectedStudy,
+    { noJobMessage }: { noJobMessage: string },
+): Promise<GuardSuccess | React.ReactElement> {
+    if (!isSubmittedStudy(study)) {
+        return <AlertNotFound title="Study was not found" message="No such study exists" />
+    }
+
+    const job = await latestSubmittedJobForStudy(study.id)
+    if (!job) {
+        return <AlertNotFound title="No submission found" message={noJobMessage} />
+    }
+
+    const stage = currentExecutionStage(job.statusChanges)
+    if (!stage) {
+        return (
+            <AlertNotFound
+                title="Outputs not yet available"
+                message="Code has been approved but execution has not started yet. Please check back shortly."
+            />
+        )
+    }
+
+    return { job, stage }
+}
