+++ src/app/[orgSlug]/study/[studyId]/_screens/reviewer-outputs-pending-screen.tsx
@@ -1,37 +1,19 @@
-import { isSubmittedStudy } from '@/schema/study'
-import { AlertNotFound } from '@/components/errors'
 import { Routes } from '@/lib/routes'
-import { currentExecutionStage } from '@/lib/study-job-status'
-import { latestSubmittedJobForStudy } from '@/server/db/queries'
 import { SecondaryAnalysisView } from '../review/secondary-analysis-view'
+import { guardExecutionStage } from './execution-stage-guard'
 import type { ScreenComponentProps } from './types'
 
 export async function ReviewerOutputsPendingScreen({
     study,
     orgSlug,
 }: Pick<ScreenComponentProps, 'study' | 'orgSlug'>) {
-    if (!isSubmittedStudy(study)) {
-        return <AlertNotFound title="Study was not found" message="No such study exists" />
-    }
+    const result = await guardExecutionStage(study, { noJobMessage: 'This study has no submitted code to review.' })
+    if (!('job' in result)) return result
 
-    const job = await latestSubmittedJobForStudy(study.id)
-    if (!job) {
-        return <AlertNotFound title="No submission found" message="This study has no submitted code to review." />
-    }
-
-    const stage = currentExecutionStage(job.statusChanges)
-    if (!stage) {
-        return (
-            <AlertNotFound
-                title="Outputs not yet available"
-                message="Code has been approved but execution has not started yet. Please check back shortly."
-            />
-        )
-    }
-
+    const { stage } = result
     return (
         <SecondaryAnalysisView
-            studyTitle={study.title}
+            studyTitle={study.title!}
             stageStatus={stage.status}
             stageStartedAt={stage.startedAt}
             previousHref={Routes.studyReviewCode({ orgSlug, studyId: study.id })}
